This patch adds an AVX-512 AESNI implementation for XTS. It is broken up into two implementations[1], similar to how AVX-512 support for RSA 2k/3k/4k signing was contributed. When a hw cipher is chosen for `x86_64`, and  the cpu has support for AVX-512 AESNI / VAES (determined by `aesni_xts_avx512_eligible`), then these new implementations are used, otherwise it falls back on the extant AESNI implementations. Below you will find the performance difference using `openssl speed` on an SPR-based system.

## AES-XTS-128

| Block Size (kb/sec) | `master`   | `xts-avx512` | % change |
| ------------------- | ---------- | ------------ | -------- |
| 16                  | 757398.48  | 1001989.34   | 32.29%   |
| 64                  | 2533507.03 | 3446985      | 36.06%   |
| 256                 | 4582908.93 | 7152731.39   | 56.07%   |
| 1024                | 6235720.36 | 11318858.07  | 81.52%   |
| 8192                | 7011349.23 | 14054834.18  | 100.46%  |
| 16384               | 7035016.53 | 14297737.9   | 103.24%  |

## AES-XTS-256

| Block Size (kb/sec) | `master`   | `xts-avx512` | % change |
| ------------------- | ---------- | ------------ | -------- |
| 16                  | 592621.28  | 851779.66    | 43.73%   |
| 64                  | 1838588.02 | 3025581.65   | 64.56%   |
| 256                 | 3674814.81 | 4589624.83   | 24.89%   |
| 1024                | 4982117.03 | 8907556.52   | 78.79%   |
| 8192                | 5529761.11 | 11516059.65  | 108.26%  |
| 16384               | 5600280.58 | 11756645.03  | 109.93%  |

---

**EDIT:**

1. @pablodelara and I spoke offline and he suggested I merge the two implementations and generate the extra rounds when generating the AES-256 implementation. The most recent [rebase](https://github.com/openssl/openssl/pull/26410/commits/c359b5e206f8072771225dfbf02e13d54743a5c5) does that.