@torben-hansen 2025-01-25
This implementation looks equal to the one that was contributed to [AWS-LC](https://github.com/aws/aws-lc) some time ago. Please note that following observation about the implementation: https://github.com/aws/aws-lc/pull/2140