@pittma 2025-01-27
> This implementation looks equal to the one that was contributed to [AWS-LC](https://github.com/aws/aws-lc) some time ago. Please note that following observation about the implementation: [aws/aws-lc#2140](https://github.com/aws/aws-lc/pull/2140)

Yes, that's right, this one has some improvements for smaller block sizes for which we saw performance regressions on via `openssl speed` that were missed via AWS-LC's/Boring's `bssl` tool, I believe, because it does not cover these smaller sizes. I have not yet added these changes to the implementation that was contributed to AWS-LC. Additionally, this implementation does not copy keys to the stack—it's based off the implementation which currently is in [an open PR](https://github.com/aws/aws-lc/pull/1550) which I've asked to be put on hold while we figured out the small block size issues (which _are_ fixed in this implementation).

I will definitely include your change, thank you for catching that and letting me know.