@Sashan 2025-02-08
I got access to modern intel CPU. I'm running `openssl speed -evp AES-256-XTS` I added breakpoints in gdb to check wether program executes functions delivered in this PR:
```
(gdb) i b 
Num     Type           Disp Enb Address            What
1       breakpoint     keep y   0x00005555555a0ef0 <main>
        breakpoint already hit 1 time
2       breakpoint     keep y   0x00007ffff78d9ca8 <aesni_xts_128_encrypt_avx512+8>
3       breakpoint     keep y   0x00007ffff78db248 <aesni_xts_128_decrypt_avx512+8>
4       breakpoint     keep y   0x00007ffff78de0a8 <aesni_xts_256_encrypt_avx512+8>
5       breakpoint     keep y   0x00007ffff78dfaa8 <aesni_xts_256_decrypt_avx512+8>
```
Unfortunately breakpoints were not hit during the execution. This is the information about CPU in system I'm using for testing:
```
processor       : 0
vendor_id       : GenuineIntel
cpu family      : 6
model           : 175
model name      : Intel(R) Xeon(R) 6780E
stepping        : 3
microcode       : 0x130001e0
cpu MHz         : 800.113
cache size      : 110592 KB
physical id     : 0
siblings        : 144
core id         : 0
cpu cores       : 144
apicid          : 0
initial apicid  : 0
fpu             : yes
fpu_exception   : yes
cpuid level     : 35
wp              : yes
flags           : fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf tsc_known_freq pni pclmulqdq dtes64 monitor ds_cpl vmx smx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid dca sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb cat_l3 cat_l2 cdp_l3 intel_ppin cdp_l2 ssbd mba ibrs ibpb stibp ibrs_enhanced tpr_shadow flexpriority ept vpid ept_ad fsgsbase tsc_adjust bmi1 avx2 smep bmi2 erms invpcid cqm rdt_a rdseed adx smap clflushopt clwb intel_pt sha_ni xsaveopt xsavec xgetbv1 xsaves cqm_llc cqm_occup_llc cqm_mbm_total cqm_mbm_local split_lock_detect user_shstk avx_vnni lam wbnoinvd dtherm ida arat pln pts hwp hwp_act_window hwp_epp hwp_pkg_req vnmi umip pku ospke waitpkg gfni vaes vpclmulqdq tme rdpid bus_lock_detect cldemote movdiri movdir64b enqcmd fsrm md_clear serialize pconfig arch_lbr ibt flush_l1d arch_capabilities
vmx flags       : vnmi preemption_timer posted_intr invvpid ept_x_only ept_ad ept_1gb flexpriority apicv tsc_offset vtpr mtf vapic ept vpid unrestricted_guest vapic_reg vid ple shadow_vmcs pml ept_mode_based_exec tsc_scaling usr_wait_pause notify_vm_exiting ipi_virt
bugs            : spectre_v1 spectre_v2 spec_store_bypass swapgs bhi
bogomips        : 4400.00
clflush size    : 64
cache_alignment : 64
address sizes   : 52 bits physical, 48 bits virtual
power management:
```

Do you have any tip for which CPU/system I should try to get to see your code in action? thanks a lot.