@Sashan 2025-02-08
this is the output of `openssl version -a` I have:
```
 openssl version -a
OpenSSL 3.5.0-dev  (Library: OpenSSL 3.5.0-dev )
built on: Sat Feb  8 15:59:22 2025 UTC
platform: linux-x86_64
options:  bn(64,64)
compiler: gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_BUILDING_OPENSSL -DNDEBUG
OPENSSLDIR: "/home/sdp/git/openssl.binaries/openssl-xts-avx512/ssl"
ENGINESDIR: "/home/sdp/git/openssl.binaries/openssl-xts-avx512/lib64/engines-3"
MODULESDIR: "/home/sdp/git/openssl.binaries/openssl-xts-avx512/lib64/ossl-modules"
Seeding source: os-specific
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffefebffff:0xfb4027bc239cb7ef:0x4c8809d0fc1c4432:0x0000000101060030:0x0000000000000000
```